+++ grid_esp/main/grid_esp32_module_ef44.c
@@ -10,7 +10,7 @@
 
 static const char *TAG = "esp32_adc";
 
-static uint32_t last_real_time[16] = {0};
+static uint64_t last_real_time[16] = {0};
 
 static uint8_t DRAM_ATTR multiplexer_index = 0;
 
@@ -66,8 +66,8 @@ static uint8_t UI_SPI_RX_BUFFER_ACTUAL[14] = {0};
 
 static volatile uint8_t UI_SPI_RX_BUFFER_LAST[16] = {0};
 
-static uint32_t encoder_last_real_time[16] = {0};
-static uint32_t button_last_real_time[16] = {0};
+static uint64_t encoder_last_real_time[16] = {0};
+static uint64_t button_last_real_time[16] = {0};
 
 static uint8_t phase_change_lock_array[16] = {0};
