+++ grid_make/grid/grid_module_ef44.c
@@ -4,7 +4,7 @@
 static volatile uint8_t adc_complete_count = 0;
 static volatile uint8_t multiplexer_index = 0;
 
-static uint32_t last_real_time[4] = {0};
+static uint64_t last_real_time[4] = {0};
 
 
 static uint8_t UI_SPI_TX_BUFFER[14] = {0};
@@ -12,8 +12,8 @@ static uint8_t UI_SPI_RX_BUFFER[14] = {0};
 
 static volatile uint8_t UI_SPI_RX_BUFFER_LAST[16] = {0};
 
-static uint32_t encoder_last_real_time[4] = {0};
-static uint32_t button_last_real_time[4] = {0};
+static uint64_t encoder_last_real_time[4] = {0};
+static uint64_t button_last_real_time[4] = {0};
 
 static uint8_t phase_change_lock_array[4] = {0};
